rename reveal to reveal.js and remove ambient modules
