+++ types/reveal.js/.eslintrc.json
