+++ types/reveal.js/plugin/highlight/highlight/index.d.ts
@@ -1,0 +1,11 @@
+
+import Reveal = require('../../../.');
+
+/**
+ * Reveal plugin highlight
+ *
+ * @see {@link https://github.com/hakimel/reveal.js/tree/master/plugin/highlight}
+ */
+declare const RevealHighlight: Reveal.PluginFunction;
+
+export = RevealHighlight;
