+++ types/reveal.js/plugin/markdown/markdown/index.d.ts
@@ -1,0 +1,11 @@
+
+import Reveal = require('../../../.');
+
+/**
+ * Reveal plugin markdown
+ *
+ * @see {@link https://github.com/hakimel/reveal.js/tree/master/plugin/markdown}
+ */
+declare const RevealMarkdown: Reveal.PluginFunction;
+
+export = RevealMarkdown;
