+++ types/reveal.js/plugin/math/math/index.d.ts
@@ -1,0 +1,12 @@
+
+
+import Reveal = require('../../../.');
+
+/**
+ * Reveal plugin math
+ *
+ * @see {@link https://github.com/hakimel/reveal.js/tree/master/plugin/math}
+ */
+declare const RevealMath: Reveal.PluginFunction;
+
+export = RevealMath;
