+++ types/reveal.js/plugin/notes/notes/index.d.ts
@@ -1,0 +1,11 @@
+
+import Reveal = require('../../../.');
+
+/**
+ * Reveal plugin notes
+ *
+ * @see {@link https://github.com/hakimel/reveal.js/tree/master/plugin/notes}
+ */
+declare const RevealNotes: Reveal.PluginFunction;
+
+export = RevealNotes;
