+++ types/reveal.js/plugin/search/search/index.d.ts
@@ -1,0 +1,11 @@
+
+import Reveal = require('../../../.');
+
+/**
+ * Reveal plugin search
+ *
+ * @see {@link https://github.com/hakimel/reveal.js/tree/master/plugin/search}
+ */
+declare const RevealSearch: Reveal.PluginFunction;
+
+export = RevealSearch;
