+++ types/reveal.js/plugin/zoom/zoom/index.d.ts
@@ -1,0 +1,12 @@
+
+import Reveal = require('../../../.');
+
+
+/**
+ * Reveal plugin zoom
+ *
+ * @see {@link https://github.com/hakimel/reveal.js/tree/master/plugin/zoom}
+ */
+declare const RevealZoom: Reveal.PluginFunction;
+
+export = RevealZoom;
