+++ types/reveal.js/tsconfig.json
@@ -19,6 +19,6 @@
     },
     "files": [
         "index.d.ts",
-        "reveal-tests.ts"
+        "reveal.js-tests.ts"
     ]
 }
