+++ types/reveal.js/tslint.json
